Refactor: Move inline styles to reusable styles.js
